% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Cross Validation}
\usage{
cv(x, y, FUN, nfolds = 5, ...)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{FUN}{\code{function} function to optimize.}

\item{nfolds}{\code{integer(1)} number of cross validation folds.}

\item{\ldots}{further arguments passed to \code{FUN}.}
}
\value{
\code{double(1)} median AUC across all cross validation splits
}
\description{
Runs a cross validation for train and prediction function.
}
\note{
The function to optimize has to accept five arguments: xtrain, ytrain, xtest,
ytest and \ldots.
}
\examples{
.rusranger <- function(xtrain, ytrain, xtest, ytest, ...) {
    rngr <- rusranger(x = xtrain, y = ytrain, ...)
    pred <- as.numeric(predict(rngr, xtest)$predictions[, 2L])
    performance(prediction(pred, ytest), measure = "auc")@y.values[[1L]]
}
cv(iris[-5], as.numeric(iris$Species == "versicolor"), .rusranger, nfolds = 3)
}
