% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rusranger.R
\name{cv_rusranger}
\alias{cv_rusranger}
\title{Cross Validation for rusranger}
\usage{
cv_rusranger(x, y, nfolds = 5, ...)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{nfolds}{\code{integer(1)} number of cross validation folds.}

\item{\ldots}{further arguments passed to \code{\link[=rusranger]{rusranger()}}.}
}
\value{
\code{double(1)} median AUC across all cross validation splits
}
\description{
Runs a cross validation for \code{\link[=rusranger]{rusranger()}}.
The cross validation optimises the AUC.
}
