% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{nested_gridsearch}
\alias{nested_gridsearch}
\title{Nested Cross Validation for Hyperparameter Search}
\usage{
nested_gridsearch(
  x,
  y,
  searchspace,
  FUN,
  nouterfolds = 5,
  ninnerfolds = 5,
  nrepcv = 2,
  ...
)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{searchspace}{\code{data.frame}, hyperparameters to tune. Column names have
to match the argument names of \code{FUN}.}

\item{FUN}{\code{function} function to optimize.}

\item{nouterfolds}{\code{integer(1)}, number of outer cross validation folds.}

\item{ninnerfolds}{\code{integer(1)}, number of inner cross validation folds.}

\item{nrepcv}{\code{integer(1)}, number repeats of inner cross validations.}

\item{\ldots}{further arguments passed to \code{\link[=gs_rusranger]{gs_rusranger()}}.}
}
\value{
\code{list}, with an element per \code{nouterfolds} containing the following
subelements:
\itemize{
\item indextrain index of the used training items.
\item indextest index of the used test items.
\item performance resulting performance (AUC).
\item selectedparams select hyperparameters.
\item gridsearch \code{data.frame}, results of the grid search.
\item nouterfolds \code{integer(1)}.
\item ninnerfolds \code{integer(1)}.
\item nrepcv \code{integer(1)}.
}
}
\description{
Run a grid search in a nested cross validation.
}
\note{
The reported performance could slightly differ from the median performance
in the reported gridsearch. After the gridsearch \code{FUN} is trained again
with the best hyperparameters which results in a new subsampling.
}
\examples{
set.seed(20220324)
iris <- subset(iris, Species != "setosa")
searchspace <- expand.grid(
   mtry = c(2, 3),
   num.trees = c(500, 1000)
)
## n(outer|inner) folds and nrepcv are too low for real world applications,
## and are just used for demonstration and to keep the run time of the examples
## low
nrcv_rusranger(
    iris[-5], as.numeric(iris$Species == "versicolor"),
    searchspace = searchspace, nouterfolds = 3, ninnerfolds = 3, nrepcv = 1
)
}
