% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{rcv}
\alias{rcv}
\title{Repeated Cross Validation}
\usage{
rcv(x, y, nfolds = 5, nrepcv = 2, FUN, ...)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{nfolds}{\code{integer(1)} number of cross validation folds.}

\item{nrepcv}{\code{integer(1)} number of repeats.}

\item{FUN}{\code{function} function to optimize.}

\item{\ldots}{further arguments passed to \code{FUN}.}
}
\value{
\code{double(5)}, minimal, 25 \% quartiel, median, 75 \% quartile and
maximal results across the repeated cross validations.
}
\description{
Runs a repeated cross validation for a function.
See also \code{\link[=cv]{cv()}}.
}
