% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rusranger.R
\name{rcv_rusranger}
\alias{rcv_rusranger}
\title{Repeated Cross Validation for rusranger}
\usage{
rcv_rusranger(x, y, nfolds = 5, nrepcv = 2, ...)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{nfolds}{\code{integer(1)} number of cross validation folds.}

\item{nrepcv}{\code{integer(1)} number of repeats.}

\item{\ldots}{further arguments passed to \code{\link[=cv_rusranger]{cv_rusranger()}}.}
}
\value{
\code{double(5)}, minimal, 25 \% quartiel, median, 75 \% quartile and
maximal AUC across the repeated cross validations.
}
\description{
Runs a repeated cross validation for \code{\link[=rusranger]{rusranger()}}.
The cross validation optimises the AUC.
}
\examples{
iris <- subset(iris, Species != "setosa")
rcv_rusranger(
    iris[-5], as.numeric(iris$Species == "versicolor"),
    nfolds = 3, nrepcv = 3
)
}
