% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusranger-package.R
\docType{package}
\name{rusranger-package}
\alias{rusranger-package}
\alias{_PACKAGE}
\title{rusranger: Modified ranger implementation to support random-under-sampling}
\description{
The random forest implementation of the ranger package is modified to support random-under-sampling. Additional helper functions for (nested)-cross-validation and diagnositc plots are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ampel-leipzig/rusranger}
  \item Report bugs at \url{https://github.com/ampel-leipzig/rusranger/issues/}
}

}
\author{
\strong{Maintainer}: Sebastian Gibb \email{mail@sebastiangibb.de} (\href{https://orcid.org/0000-0001-7406-4443}{ORCID})

Authors:
\itemize{
  \item Paul Ahrens \email{paul.ahrens@medizin.uni-leipzig.de}
  \item Daniel Steinbach \email{daniel.steinbach@medizin.uni-leipzig.de}
}

}
\keyword{internal}
