% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusranger.R
\name{rusranger}
\alias{rusranger}
\title{RUS ranger}
\usage{
rusranger(
  x,
  y,
  probability = TRUE,
  classification = !probability,
  min.node.size = if (probability) 10 else 1,
  replace = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{matrix}/\code{data.frame}, feature matrix, see \code{\link[=ranger]{ranger()}} for
details.}

\item{y}{\code{numeric}/\code{factor}, classification labels, see \code{\link[=ranger]{ranger()}} for
details.}

\item{probability}{\code{logical(1)}, grow probability trees, see \code{\link[=ranger]{ranger()}}
for details.}

\item{classification}{\code{logical(1)}, run classification even if \code{y} is
\code{numeric}, see \code{\link[=ranger]{ranger()}} for details.}

\item{min.node.size, }{same as in \code{\link[=ranger]{ranger()}}}

\item{replace, }{subsampling without (default, \code{replace = FALSE}) or with
resampling, see \code{\link[=ranger]{ranger()}} for details.}

\item{\ldots}{further arguments passed to \code{\link[=ranger]{ranger()}}.}
}
\value{
\code{ranger} object, see \code{\link[=ranger]{ranger()}} for details.
}
\description{
Adapted default settings to the \code{\link[=ranger]{ranger()}} to support
random-under-sampling. Additionally the default settings are modified to the
most common classification settings used in the AMPEL project.
}
\details{
In contrast to \code{\link[=ranger]{ranger()}} \code{rusranger()} currently just supports binary
classifications.
}
\references{
\emph{AMPEL} project:
Analysis and Reporting System for the Improvement of Patient Safety
through Real-Time Integration of Laboratory Findings,
https://ampel.care.
}
\seealso{
\code{\link[=ranger]{ranger()}}
}
