% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smote.R
\name{smote}
\alias{smote}
\title{SMOTE}
\usage{
smote(x, y, k = 5L, ndups = 0L)
}
\arguments{
\item{x}{\code{matrix}, features}

\item{y}{\code{factor}, class (length has to be equal to number of rows in \code{x}).}

\item{k}{\code{numeric(1)}, number of nearest neighbours.}

\item{ndups}{\code{numeric(1)}, number of synthetic duplicates to upsample
minority. Default is \code{0} which tries to upsample to majority.}
}
\value{
\code{list} with elements \code{x} features with SMOTEd samples and \code{y} with
added minority class elements.
}
\description{
SMOTE: Synthetic Minority Over-sampling Technique.
}
\references{
Chawla, Bowyer, Hall and Kegelmeyer. 2002.
SMOTE: Synthetic Minority Over-sampling Technique.
Journal of Artificial Intelligence Research 16. 321–357.
\doi{10.1613/jair.953}
}
